Add private headers to share platform specific functions
Rework the compilation to build the correct objects for each target.
